+++ aucom/src/main/java/org/bham/aucom/diagnoser/t2gram/detector/anomalyclassificator/StatisticalHysterisisAnomalyClassificator.java
@@ -12,8 +12,8 @@ public class StatisticalHysterisisAnomalyClassificator extends AbstractAnomalyCl
 
     public StatisticalHysterisisAnomalyClassificator(Double mean, Double variance) {
         super("StatisticalHysterisisThreshold");
-        this.mean = mean.doubleValue();
-        this.variance = variance.doubleValue();
+        this.mean = mean;
+        this.variance = variance;
     }
 
     @Override
